Update inbox state to include installation timestamps
